### 2024-07-03 #6959 @editorialbot
[PRE REVIEW]: BrightEyes-MCS: a control suite for multichannel scanning microscopy
**Submitting author:** @VicidominiLab (<a href="http://orcid.org/0000-0003-0026-747X">Mattia Donato</a>)
**Repository:** https://github.com/VicidominiLab/BrightEyes-MCS
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.9.0
