### 2024-07-03 8060bbf
Added fastapi to the requirements.txt
