### 2024-07-03 fa7c92f
Main without paper
