+++ paper.md
@@ -1,236 +1,0 @@
----
-title: 'BrightEyes-MCS: a control software for multichannel scanning microscopy'
-tags:
-  - Image Scanning Microscopy
-  - Fluorescence Microscopy
-  - Super-Resolution Microscopy
-  - Scanning Microscopy
-  - Microscope automation
-  
-authors:
-  - name: Mattia Donato
-    orcid: 0000-0003-0026-747X
-    #equal-contrib: true
-    affiliation: "1" # (Multiple affiliations must be quoted)
-  - name: Eli Slenders
-    #orcid:
-    #equal-contrib: false
-    affiliation: "1" # (Multiple affiliations must be quoted)
-  - name: Alessandro Zunino
-    orcid: 0000-0002-2512-8751
-    #orcid:
-    #equal-contrib: false
-    affiliation: "1" # (Multiple affiliations must be quoted)
-  - name: Luca Bega
-    #equal-contrib: false
-    affiliation: "1" # (Multiple affiliations must be quoted)
-  - name: Giuseppe Vicidomini
-    orcid: 0000-0002-3085-730X
-    #equal-contrib: false
-    corresponding: true
-    affiliation: "1" # (Multiple affiliations must be quoted)
-
-#  - name: Author Without ORCID
-#    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
-#    affiliation: 2
-#  - name: Author with no affiliation
-#    corresponding: true # (This is how to denote the corresponding author)
-#    affiliation: 3
-#  - given-names: Ludwig
-#    dropping-particle: van
-#    surname: Beethoven
-#    affiliation: 3
-affiliations:
- - name: Molecular Microscopy and Spectroscopy, Istituto Italiano di Tecnologia, Genoa, 16152, Italy
-   index: 1
-# - name: Institution Name, Country
-#   index: 2
-# - name: Independent Researcher, Country
-#   index: 3
-date: 13 August 2017
-bibliography: paper.bib
-
-# Optional fields if submitting to a AAS journal too, see this blog post:
-# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-# aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-# aas-journal: Astrophysical Journal <- The name of the AAS journal.
----
-
-
-# Introduction
-
-Fluorescence microscopy is an essential workhorse of biomedical sciences, thanks to its capabilities to provide specific and quantitative information of the observed specimens. The advent of super-resolution techniques has further improved the quality of the images produced by optical microscopes. Among the  numerous super-resolution techniques, image scanning microscopy (ISM)  emerged as a robust and reliable technique, being able to provide gentle imaging at high signal-to-noise ratio (SNR) with excellent optical sectioning​ [@Castello1; @Tortarolo1; @Perego1]​.  An ISM microscope shares the same architecture of a confocal laser scanning microscope (CLSM) but  exploits a pixeled detector. Each detector  element acts as a displaced  small pinhole and generates a confocal-like image after a full scan of the field of view.  As the detector is made by a matrix of pixel, the raw ISM dataset is not a 2D image, but it is a 4D dataset, which can be seen in two ways: for each single pixel scanned on the sample plane is associated a micro-image of the detector; vice-versa, for each physical pixel on the detector plane is associated a full-scanned field of view on the sample plane [@Sheppard1; @Muller1]​.  Using tailored reconstruction algorithms, the images of the raw ISM dataset can be fused to  produce a single super-resolution image using all the photons collected by the detector, guaranteeing an excellent SNR​ [@Zunino1; @Zunino2]​.
-
[…]
