+++ paper.bib
@@ -1,168 +1,0 @@
-@article{Castello1,
-  author    = {M. Castello and G. Tortarolo and M. Buttafava and T. Deguchi and F. Villa and S. Koho and L. Pesce and M. Oneto and S. Pelicci and L. Lanzanó and P. Bianchini and C. J. R. Sheppard and A. Diaspro and A. Tosi and G. Vicidomini},
-  title     = {A robust and versatile platform for image scanning microscopy enabling super-resolution FLIM},
-  journal   = {Nature Methods},
-  volume    = {16},
-  pages     = {175--178},
-  year      = {2019},
-  doi       = {10.1038/s41592-018-0291-7},
-  url       = {https://www.nature.com/articles/s41592-018-0291-7}
-}
-
-@article{Tortarolo1,
-  author    = {G. Tortarolo and A. Zunino and F. Fersini and M. Castello and S. Piazza and C. J. R. Sheppard and P. Bianchini and A. Diaspro and S. Koho and G. Vicidomini},
-  title     = {Focus image scanning microscopy for sharp and gentle super-resolved microscopy},
-  journal   = {Nature Communications},
-  volume    = {13},
-  year      = {2022},
-  doi       = {10.1038/s41467-022-31408-6},
-  url       = {https://www.nature.com/articles/s41467-022-31408-6}
-}
-
-@article{Perego1,
-  author    = {E. Perego and S. Zappone and F. Castagnetti and D. Mariani and E. Vitiello and J. Rupert and E. Zacco and G. G. Tartaglia and I. Bozzoni and E. Slenders and G. Vicidomini},
-  title     = {Single-photon microscopy to study biomolecular condensates},
-  journal   = {Nature Communications},
-  volume    = {14},
-  pages     = {8224},
-  year      = {2023},
-  doi       = {10.1038/s41467-023-15070-4},
-  url       = {https://www.nature.com/articles/s41467-023-15070-4}
-}
-
-@article{Sheppard1,
-  author    = {C. Sheppard},
-  title     = {Super-resolution in confocal imaging},
-  journal   = {Optik},
-  volume    = {80},
-  pages     = {53--54},
-  year      = {1988}
-}
-
-@article{Muller1,
-  author    = {C. B. Müller and J. Enderlein},
-  title     = {Image scanning microscopy},
-  journal   = {Physical Review Letters},
-  volume    = {104},
-  year      = {2010},
-  doi       = {10.1103/PhysRevLett.104.198101},
-  url       = {https://journals.aps.org/prl/abstract/10.1103/PhysRevLett.104.198101}
-}
-
-@article{Zunino1,
-  author    = {A. Zunino and M. Castello and G. Vicidomini},
-  title     = {Reconstructing the Image Scanning Microscopy Dataset: an Inverse Problem},
-  journal   = {Journal of Optics},
-  volume    = {39},
-  issue     = {6},
-  pages     = {064004},
-  year      = {2022},
-  doi       = {10.1088/1361-6420/accdc5},
-  url       = {https://iopscience.iop.org/article/10.1088/1361-6420/accdc5}
-}
-
-@article{Zunino2,
-  author    = {A. Zunino and E. Slenders and F. Fersini and A. Bucci and M. Donato and G. Vicidomini},
-  title     = {Open-source tools enable accessible and advanced image scanning microscopy data analysis},
-  journal   = {Nature Photonics},
-  volume    = {17},
-  pages     = {457--458},
-  year      = {2023},
-  doi       = {10.1038/s41566-023-01112-y},
-  url       = {https://www.nature.com/articles/s41566-023-01112-y}
-}
-
-@article{Buttafava1,
[…]
