### 2024-07-18 #6959 @Kevin-Mattheus-Moerman
Finally I see this work features hardware requirements. I need to point out that although we have successfully reviewed/published software works that link with specialised hardware in the past, we have also seen severe difficulties in finding reviewers or completing the reviews fully which in some cases has led to the papers ending up retracted. So, since this work appears to require particular NI FPGA's, I just want to add a word of warning that the finding reviewers may take long or be un-succesfull altogether, and that the review process may take longer than usual. The success strongly depends on how clear and well documented your package is in terms of what the hardware requirements are and how to set everything up for testing. If you feel these aspects are in good shape that is great, if not, then pushing now for these to be extended would be a good idea. Further more, when asked by the editor, it would be really helpful if you could help suggest any potential reviewers (provided there is no conflict of interest).
So in short, hardware related submissions in JOSS are possible, however they can be tricky, so keep that in mind. Thanks.
