### 2024-08-08 #6959 @VicidominiLab
Dear @Kevin-Mattheus-Moerman,

Thank you for your feedback on our submission and your patience. Thank you for the comment about  the bibliography issues and the contributing guidelines. In the next commit I will fix the .bib file, the contributing.md and I will try to push more clearly hardware requirements and setup process.

We chose JOSS for our submission because it aligns with our project goals and audience. We've noticed similar projects microscopy-related, like [ImSwitch](https://joss.theoj.org/papers/10.21105/joss.03394), which also involve hardware components. I understand the complexities involved in reviewing hardware-dependent software but in scientific-paper revision it is rare that the reviewer has the hardware/complete system in hand. In our case, while the NIFPGA hardware mentioned in the paper is essential for running the BrightEyes-MCS acquisition, the additional hardware components (such as detectors, optics, optoelectronics, lasers, etc...) are not strictly required for testing the software.

Regarding the review process, I understand the potential challenges in finding suitable reviewers due to the hardware aspect. To assist, I would like to suggest two potential reviewers who they should have a system similar to ours.

- J. Abramo (GitHub: [jacopoabramo](https://github.com/jacopoabramo)) - the current maintainer of ImSwitch.
- B. Diederich (GitHub: [beniroquai](https://github.com/beniroquai)) - the founder of OpenUC2.

Thank you once again for your assistance. Please let me know if there's anything else I can do to support the process.

Kind regards,

Mattia
