### 2024-08-19 #6959 @Kevin-Mattheus-Moerman
@adamltyson thanks! Please check all my comments above. Some work is still needed to proceed and I did raise the issue of hardware requirements. But, if we can find reviewers that can help, then I'm happy to proceed to review.
