### 2024-08-19 #6959 @adamltyson
Hi @VicidominiLab, I'll take over the submission from now on.

As @Kevin-Mattheus-Moerman mentioned, it may be tricky, or even not possible for this to be reviewed, but we'll try our best! I will push ahead in finding reviewers, but in the mean time, can you make sure to address @Kevin-Mattheus-Moerman's comments about the references and the contributing guide?

Any questions, please let me know.
