### 2024-08-21 #6959 @jacopoabramo
Greetings @adamltyson, sorry for the late reply - I was on vacation until yesterday.

I'd be happy to review the paper. Since it's my first time reviewing for JOSS do you have any input on where I should start with the review? Any tip is appreciated. Thank you in advance.
