### 2024-08-21 #7125 @adamltyson
@VicidominiLab, while the reviewers start, there are a few outstanding issues raised by @Kevin-Mattheus-Moerman to sort out. Could you:
- Check the reference check above, and address the reported DOI issues. You can run `@editorialbot check references` at any time to re-check.
- Ensure the repository has a contributing guide. This can be as simple as adding a `CONTRIBUTING.md` file to the repo (see [here](https://contributing.md/example/) for examples).

Could you also:
- Make sure the `LICENSE.md` contains  [OSI approved license](https://opensource.org/licenses) text, and only that. I.e., can you copy and paste a license, without adding any other text. Any other information can be added to the readme, or anywhere else in the repository.
- Make sure the paper has a clear "Statement of need" section. See the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) for more information.
- Run the paper through a spelling and grammar checker. There are a few small issues in the paper text.

Any questions, please let me know.
