### 2024-08-22 #7125 @adamltyson
Hi @VicidominiLab, I noticed that the paper is also considerably over the [recommended 250-1000 words](https://joss.readthedocs.io/en/latest/paper.html). Would you be able to cut down the length of the text? A lot of the technical details about the software could probably be added to the documentation instead.
