### 2024-08-22 #7125 @jacopoabramo
> Hi @VicidominiLab, I noticed that the paper is also considerably over the [recommended 250-1000 words](https://joss.readthedocs.io/en/latest/paper.html). Would you be able to cut down the length of the text? A lot of the technical details about the software could probably be added to the documentation instead.

I believe that the GUI section could be removed entirely and ported to the documentation in this sense. The other chapters describing the hardware and software architecture are still relevant to the paper, while the description of the GUI fits more into the user manual - with proper details on how to use it eventually.
