### 2024-08-22 #7125 @raacampbell
@adamltyson I assume we should be making relevant comments regarding things like clarity of writing for specific sections, choice of wording, etc. Those are issues related to the paper itself rather than the software. Should we be raising those issues here alone or should also raise these non-code points as issues in the repo?
