### 2024-08-22 #7125 @raacampbell
What is the preferred way to signal when such issues are resolved? Do you want checkboxes in the text or emoticons?
