### 2024-08-22 #7125 @raacampbell
@VicidominiLab, it would be helpful to sort out the [broken DOI links](https://github.com/openjournals/joss-reviews/issues/7125#issuecomment-2301582178) ASAP since copy/paste from the PDF works poorly due to the watermark text.
