### 2024-08-22 #7125 @adamltyson
FYI @raacampbell, the paper (`paper.md`) and references (`paper.bib`) can be found in [this branch](https://github.com/VicidominiLab/BrightEyes-MCS/tree/paper), if you want access to the underlying files. You're also welcome to raise a PR for small changes if it's easier than posting them here.
