### 2024-08-22 #7125 @raacampbell
@VicidominiLab, I think the title ("_BrightEyes-MCS: a control software for multichannel scanning microscopy_") perhaps does not accurately reflect what the paper is about and what the software is for. Perhaps the scope of the project is not sufficiently clear from the open sentences.

* [ ] When I read the title, I expect you have written general-purpose scanning software for fluorescence microscopy. However, as I read the paper and look at your previous work, it seems clear the main goal of the software is to achieve open-source ISM with SPAD arrays. That said, I see from Figure 2 and the Conclusion that your software will accept PMT inputs. The opening paragraph and title should make the scope of the work clearer.
* [ ] I would have found it helpful if it was mentioned in the first paragraph that NSPARC and Airyscan are commercial ISM implementations. I of course had heard of those before but had not heard the acronym "ISM".
