### 2024-08-22 #7125 @raacampbell
@VicidominiLab, the writing is clear but the text does need to be proof-read by a native-proficiency English speaker: there are typos and grammatical errors throughout and these detract from your nice work. e.g. line 89 `photons multiplier tubes` should read `photomultiplier tubes`.
