### 2024-08-22 #7125 @raacampbell
@VicidominiLab, the software is presumably for linear (galvo) scanners only? i.e. does not handle resonant scanners.

* [ ] Please clarify this explicitly in your text.
