+++ README.md
@@ -105,8 +105,9 @@ BrightEyes-MCS supports only FPGA from NI. At the moment the bitfile are built f
 | 1    | NI FPGA PXIe-7856 (Single-board)                   | NI Chassis + NI Thunderbolt | Full supported, currently in use in our lab |
 | 2    | NI FPGA USB-7856R (Single-board)                   | -                           | Full supported, currently in use in our lab |
 | 3    | NI FPGA USB-7856R OEM (Single-board)               | -                           | Should work but not fully tested            |
-| 4    | NI FPGA PXIe-7822 + NI FPGA PXIe-7856 (Dual-board) | NI Chassis + NI Thunderbolt | Should work but not fully tested            |
-| 5    | NI FPGA PXIe-7820R (Dual-board)                    | No Analog integrated        | Should work but no tested                   |
+| 4    | NI FPGA PXIe-7822 + NI FPGA PXIe-7856 (Dual-board) | NI Chassis + NI Thunderbolt | Full supported, currently in use in our lab |
+| 4    | NI FPGA PXIe-7822 (Singel-board)                   |  NI Chassis + NI Thunderbolt +  External DAC required       | Full supported, currently in use in our lab |
+| 5    | NI FPGA PCIe-7820R                                 | External DAC required       | Full supported, currently in use in our lab |
 
 **Warning!! To avoid any damage of your equipment, please verify that the pinout described in I/O Table are compatible with your actual system.**
 
@@ -127,6 +128,15 @@ PyCharm provides a easy graphical interfaces to donwload the source code, to cre
 
 
 
+## Firmware
+**IMPORTANT**
+The firmwares needed for running the NI FPGA are not included in the BrightEyes-MCS tree.
+It is present during the installation phase a graphical tool to download and extract them.
+In case of issues it is possible download directly from the repository of [BrightEyes-MCSLL](https://github.com/VicidominiLab/BrightEyes-MCSLL).
+
+## I/O table
+The I/O table depeands by the firmware. The tables are available on the documentations of [BrightEyes-MCSLL](https://github.com/VicidominiLab/BrightEyes-MCSLL)
+
 # Getting started
 
 
@@ -141,12 +151,6 @@ install.bat
 and following the instructions on the terminal.
 
 Once installed you will have the link with icon of BrightEyes-MCS and on your Desktop.
-
-## Firmware
-**IMPORTANT**
-The firmwares needed for running the NI FPGA are not included in the BrightEyes-MCS tree.
-It is present during the installation phase a graphical tool to download and extract them.
-In case of issues it is possible download directly from the repository of [BrightEyes-MCSLL](https://github.com/VicidominiLab/BrightEyes-MCSLL).
 
 ---
 
@@ -208,308 +212,4 @@ Compile the Cython modules:
 Run BrightEyes-MCS:
 
 ```run-brighteys-mcs.bat```
-
-
-# Hardware:
-Here below a typical connection scheme of BrightEyes-MCS.
-
-
-
-## Microscope:
-Galvo, laser, etc etc... TO-Domain
-
-
-# BrightEyes-MCS-LowLevel Firmwares
-
-## I/O Table
-
-|                                           | I/O | USB-7856R (Single-Board) | PXIe-7856R (Single-Board) |
-|-------------------------------------------|-----|----------------------|-----------------------|
-| SPAD CHANNEL 0                            | IN  | Connector0/DIO0      | Connector1/DIO0       |
-| ...                                       | ... | ..                   | ...                   |
-| SPAD CHANNEL 15                           | IN  | Connector0/DIO15     | Connector1/DIO15      |
-| SPAD CHANNEL 16                           | IN  | Connector0/DIO16     | Connector1/DIO16      |
[…]
