### 2024-09-03 #7125 @beniroquai
Quick question @VicidominiLab, is the software inteded be used on e.g. MAC or Linux?
