### 2024-09-06 #7125 @VicidominiLab
> Quick question @VicidominiLab, is the software inteded be used on e.g. MAC or Linux?

Sorry I read only today. BrightEyes-MCS supports only Windows at the moment. I will add it in the requirements.
