### 2024-09-10 #7125 @beniroquai
Thanks @VicidominiLab, I was able to get hold on a windows machine now. I follow your instructions and fail here:
![image](https://github.com/user-attachments/assets/e79d477f-62b7-4e22-a9dd-5a1b35b5d0e6)

Is there any way to test the sotware without the FPGA software from NI?
