### 2024-09-10 #7125 @raacampbell
>
> Is there any way to test the sotware without the FPGA software from NI?

Along these lines: a comment I have is that it would be very helpful if there was as a documented way of starting the software with no hardware or with simulated hardware. This is useful for evaluation (e.g. this review) but is particularly useful  for software development.
