### 2024-09-10 #7125 @jacopoabramo
> Hi @raacampbell, @jacopoabramo, @beniroquai 👋
>
> Just checking in to see how the reviews are going. Is there anything I can do to help/clarify?

Sorry, I'm currently away on a workshop in Seattle until Wednesday. I will get back on the review from Thursday onward.
