### 2024-09-12 #7125 @VicidominiLab
> Is there any way to test the sotware without the FPGA software from NI?
@beniroquai thank I put the wrong link, I will correct it! Actually the NI Driver and LabView Runtime needed can be downloaded at https://www.ni.com/en/support/downloads/drivers/download.ni-r-series-multifunction-rio.html
