### 2024-09-12 #7125 @VicidominiLab
@raacampbell
The whole project is built around FPGA NI, and during the acquisition some parts are directly mapped from the FPGA FSMs status. There is a data simulator but this is also built into the FPGA firmware. I understand that it would be useful to have a way to start the acquisition with simulated FPGA hardware, but implementing such a solution would be quite complex at this time. It would essentially require the development of a separate project to create a mock version of the system.
