### 2024-09-16 #7125 @raacampbell
The `install.bat` install method failed with miniconda for me. I made an environment and tried to install in that. After double-clicking on the icon I get an error window stating that the Cython modules can not be found.

I have not there yet been able to see what happens when I try running a system with no NI hardware.
