### 2024-09-16 #7125 @raacampbell
This is relatively elaborate acquisition software. I think it would benefit from more expansive documentation. At present I see only brief install instructions in the README file. I would rather see much more detailed instructions in a GitHub wiki or some other similar suitable site.
