### 2024-09-16 #7125 @mattia-donato
> The `install.bat` install method failed with miniconda for me. I made an environment and tried to install in that. After double-clicking on the icon I get an error window stating that the Cython modules can not be found. If I try to compile this happens:
>
> ```
> (venv) (brighteyes) C:\Users\shopp\Downloads\BrightEyes-MCS-main\BrightEyes-MCS-main>python setup.py build_ext -inplace
> No parent package detected, impossible to derive `name`
> running build_ext
> error: --plat-name must be one of ('win32', 'win-amd64', 'win-arm32', 'win-arm64')
> ```
>
> I have not there yet been able to see what happens when I try running a system with no NI hardware.

@raacampbell
I suppose you are not using Visual Studio compiler but **msys2**. In this case, "python setup.py build_ext -inplace" is not enough. You need to run from a terminal (with venv or conda activated) the compile_cython_msys2.bat which set some env variables and parameter to the compilation.

**compile_cython_msys2.bat**:
```
@REM VERSION WITH MSYS2
@ECHO This is the script to compile the Cython part of the code
@ECHO "msys2" must be installed on your system in the default folder (C:\msys64)
@ECHO you can download it from https://www.msys2.org/
@ECHO and inside msys2 environment gcc must be installed
@ECHO (you can install it inside msys2 environment with the
@ECHO command 'pacman -S mingw-w64-ucrt-x86_64-gcc' inside msys2 env)

SET PATH=%PATH%;C:\msys64\usr\bin;C:\msys64\ucrt64\bin
python setup.py build_ext --inplace --force --compiler=mingw32 -DMS_WIN64

```
