### 2024-09-17 VicidominiLab/BrightEyes-MCS#2 @raacampbell
Unable to compile/install
Moving this problem from https://github.com/openjournals/joss-reviews/issues/7125 to an Issue.

The `install.bat` install method failed with miniconda for me. I made an environment and tried to install in that. After double-clicking on the icon I get an error window stating that the Cython modules can not be found. If I try to compile this happens:

```
(venv) (brighteyes) C:\Users\shopp\Downloads\BrightEyes-MCS-main\BrightEyes-MCS-main>python setup.py build_ext -inplace
No parent package detected, impossible to derive `name`
running build_ext
error: --plat-name must be one of ('win32', 'win-amd64', 'win-arm32', 'win-arm64')
```

The response from @mattia-donato was:

I suppose you are not using Visual Studio compiler but msys2. In this case, `python setup.py build_ext -inplace` is not enough. You need to run from a terminal (with venv or conda activated) the `compile_cython_msys2.bat` which set some env variables and parameter to the compilation.

```
compile_cython_msys2.bat:

@REM VERSION WITH MSYS2
@ECHO This is the script to compile the Cython part of the code
@ECHO "msys2" must be installed on your system in the default folder (C:\msys64)
@ECHO you can download it from https://www.msys2.org/
@ECHO and inside msys2 environment gcc must be installed
@ECHO (you can install it inside msys2 environment with the
@ECHO command 'pacman -S mingw-w64-ucrt-x86_64-gcc' inside msys2 env)

SET PATH=%PATH%;C:\msys64\usr\bin;C:\msys64\ucrt64\bin
python setup.py build_ext --inplace --force --compiler=mingw32 -DMS_WIN64
```

I installed [MSYS2](https://www.msys2.org/) then in the window that came up ran `pacman -S mingw-w64-ucrt-x86_64-gcc` as instructed. After this, the compile script (`compile_cython_msys2.bat`) seems to do something but ends with many errors.

```
brighteyes_mcs/libs/cython/fastconverter.c: In function '__Pyx_ImportType_3_0_11':
brighteyes_mcs/libs/cython/fastconverter.c:30252:13: warning: unknown conversion type character 'z' in format [-Wformat=]
30252 |             "%s.%s size changed, may indicate binary incompatibility. "
      |             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
brighteyes_mcs/libs/cython/fastconverter.c:30253:24: note: format string is defined here
30253 |             "Expected %zd from C header, got %zd from PyObject",
      |                        ^
brighteyes_mcs/libs/cython/fastconverter.c:30252:13: warning: unknown conversion type character 'z' in format [-Wformat=]
30252 |             "%s.%s size changed, may indicate binary incompatibility. "
      |             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
brighteyes_mcs/libs/cython/fastconverter.c:30253:47: note: format string is defined here
30253 |             "Expected %zd from C header, got %zd from PyObject",
      |                                               ^
brighteyes_mcs/libs/cython/fastconverter.c:30252:13: warning: too many arguments for format [-Wformat-extra-args]
30252 |             "%s.%s size changed, may indicate binary incompatibility. "
      |             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
error: command 'C:\\msys64\\ucrt64\\bin\\gcc.exe' failed with exit code 1

(venv) (brighteyes) C:\Users\shopp\Downloads\BrightEyes-MCS-main\BrightEyes-MCS-main>

```

I suggest much more detailed and hand-holding install instructions, as the steps are not trivial.
