### 2024-09-17 #7125 @raacampbell
Moved the install issue to: https://github.com/VicidominiLab/BrightEyes-MCS/issues/2
