### 2024-09-17 VicidominiLab/BrightEyes-MCS#2 @mattia-donato
Thank you for feedback!
I read that you use conda did you test it inside a conda env or in the system?
Actually the installation I used was using a "pure" Python and creating a venv.
It should be the same, but I worried that some packages are different.

Could you please attach here the "pip freeze" (or the equivalent for conda) in the current env?
Which version of Python are you using?
