### 2024-09-17 VicidominiLab/BrightEyes-MCS#2 @mattia-donato
@raacampbell  Thanks for your feedback!

I read that you use conda. Did you test the installation inside a conda env or in directly the system?
t should be the same, but I worried that some packages are different.
Actually the installation I used was using a "pure" Python and creating a venv from scratch without conda.

Could you please attach here the "pip freeze" (or the equivalent for conda) in the current env?
Which version of Python are you using?
