### 2024-09-17 VicidominiLab/BrightEyes-MCS#2 @mattia-donato
Actually they looks like warnings and not errors, maybe the libraries actually are compiled.

Could please you try from that folder and in the venv the following command?
`python -c "import brighteyes_mcs.libs.cython.fastconverter as fc; print(fc.convertRawDataToCountsDirect.__name__)"`
