### 2024-09-23 VicidominiLab/BrightEyes-MCS#2 @raacampbell
Sorry, I think I can't contribute more to finding a solution to this.
