+++ requirements.txt
@@ -10,6 +10,7 @@ pyqtdarktheme
 imageio
 qtconsole
 napari
+uvicorn
 importlib-metadata
 nifpga-fast-fifo-recv>=0.100.1
 #ultradict
@@ -17,4 +18,4 @@ nifpga-fast-fifo-recv>=0.100.1
 atomics
 matplotlib
 fastapi
-#brighteyes-flim  # facultative activate in case of FLIM analysis/preview
+#brighteyes-flim #facultative activate in case of FLIM analysis/preview
