### 2024-09-25 64e67d2
compile_cython.py has the possibility to use argumetns and it has MSYS2/GCC installation
