+++ README.md
@@ -15,6 +15,7 @@ Main Features
 - Support up 2 analog channels (PMT)
 - Pan & Zoom in realtime
 - Scan along each axis and XYZ multi-stack
+  - (Supported any linear voltage-controlled actuator *i.e.* galvo mirror or piezo-stage)
 - Time lapse / Macro
 - Data saved in HDF5 with metadata (current configurations, extra information, users comments etc etc).
   - The saved data can be analyzed with Napari (through Napari-ISM plugin) just by a few click
@@ -43,6 +44,71 @@ single-photon time-tagging microscopy therefore fluorescence spectroscopy, fluor
 
 ---
 
+
+---
+
+# Requirements 
+
+## Software
+- Windows OS
+- Python **3.10**
+- - Download [installer](https://www.python.org/ftp/python/3.10.8/python-3.10.8-amd64.exe)
+- - Or `winget install -e --id Python.Python.3.10`
+- Git
+- - [Git for Windows] (https://gitforwindows.org/)
+- - Or `winget install -e --id Git.Git`
+- FPGA drivers installed,
+- - NI R Series Multifunction [RIO driver](https://www.ni.com/en/support/downloads/drivers/download.ni-r-series-multifunction-rio.html)
+- A **C compiler** installer, either
+  - GCC [MSYS2](https://www.msys2.org/) (`winget install -e --id MSYS2.MSYS2`)
+  - or Microsoft Visual Studio with C++ build tools.
+
+
+## Hardware
+BrightEyes-MCS supports only FPGA from NI. At the moment the bitfile are built for the following boards.
+
+| Type | Board                                              | Extra req.                  | Tested                                      |
+|------|----------------------------------------------------|-----------------------------|---------------------------------------------|
+| 1    | NI FPGA PXIe-7856 (Single-board)                   | NI Chassis + NI Thunderbolt | Full supported, currently in use in our lab |
+| 2    | NI FPGA USB-7856R (Single-board)                   | -                           | Full supported, currently in use in our lab |
+| 3    | NI FPGA USB-7856R OEM (Single-board)               | -                           | Should work but not fully tested            |
+| 4    | NI FPGA PXIe-7822 + NI FPGA PXIe-7856 (Dual-board) | NI Chassis + NI Thunderbolt | Full supported, currently in use in our lab |
+| 4    | NI FPGA PXIe-7822 (Single-board)                   |  NI Chassis + NI Thunderbolt +  External DAC required       | Full supported, currently in use in our lab |
+| 5    | NI FPGA PCIe-7820R                                 | External DAC required       | Full supported, currently in use in our lab |
+
+**Warning!! To avoid any damage of your equipment, please verify that the pinout described in I/O Table are compatible with your actual system.**
+
+## I/O table
+The I/O table depeands by the firmware. The tables are available on the documentations of [BrightEyes-MCSLL](https://github.com/VicidominiLab/BrightEyes-MCSLL)
+
+
+
+
+
+# Getting started
+## Installation 
+
+
+There are two possibility:
+- Download the [Zip](https://github.com/VicidominiLab/BrightEyes-MCS/archive/refs/heads/main.zip) of the repository and extract it where you prefer OR
+- Clone the repo:
+- - `git clone https://github.com/VicidominiLab/BrightEyes-MCS.git`
+
+Python **3.10** and the MSYS2 or Visual Studio C++ MUST BE installed before run the installer!
+
+From the cmd.exe (NOT powershell!)
+
+```
+cd BrightEyes-MCS
+install.bat
+```
+
+and following the instructions on the terminal.
+Once installed you will have the link with icon of BrightEyes-MCS and on your Desktop.
+
+---
+
+
 # License
 
[…]
