### 2024-09-26 VicidominiLab/BrightEyes-MCS#2 @mattia-donato
Thank you for your feedback. I completly changed the installer script and finally now works also with conda.
And I updated the README about the installation and the requirements.
