### 2024-09-27 57a6a68
Added copy current conf for macro
