### 2024-09-27 da87eec
GUI improved FCS Batch tab
