+++ brighteyes_mcs/gui/main_window_design.py
@@ -41,27 +41,6 @@ class Ui_MainWindowDesign(object):
 
         self.gridLayout_4 = QGridLayout()
         self.gridLayout_4.setObjectName(u"gridLayout_4")
-        self.checkBox_autoscale_img = QCheckBox(self.tab_preview)
-        self.checkBox_autoscale_img.setObjectName(u"checkBox_autoscale_img")
-        self.checkBox_autoscale_img.setChecked(True)
-
-        self.gridLayout_4.addWidget(self.checkBox_autoscale_img, 2, 0, 1, 1)
-
-        self.checkBox_lockMove = QCheckBox(self.tab_preview)
-        self.checkBox_lockMove.setObjectName(u"checkBox_lockMove")
-
-        self.gridLayout_4.addWidget(self.checkBox_lockMove, 0, 0, 1, 1)
-
-        self.label_60 = QLabel(self.tab_preview)
-        self.label_60.setObjectName(u"label_60")
-
-        self.gridLayout_4.addWidget(self.label_60, 0, 2, 1, 1)
-
-        self.label_plot_ch = QLabel(self.tab_preview)
-        self.label_plot_ch.setObjectName(u"label_plot_ch")
-
-        self.gridLayout_4.addWidget(self.label_plot_ch, 2, 2, 1, 1)
-
         self.comboBox_view_projection = QComboBox(self.tab_preview)
         self.comboBox_view_projection.addItem("")
         self.comboBox_view_projection.addItem("")
@@ -72,6 +51,17 @@ class Ui_MainWindowDesign(object):
         self.comboBox_view_projection.setObjectName(u"comboBox_view_projection")
 
         self.gridLayout_4.addWidget(self.comboBox_view_projection, 0, 3, 1, 1)
+
+        self.checkBox_autoscale_img = QCheckBox(self.tab_preview)
+        self.checkBox_autoscale_img.setObjectName(u"checkBox_autoscale_img")
+        self.checkBox_autoscale_img.setChecked(True)
+
+        self.gridLayout_4.addWidget(self.checkBox_autoscale_img, 2, 0, 1, 1)
+
+        self.label_plot_ch = QLabel(self.tab_preview)
+        self.label_plot_ch.setObjectName(u"label_plot_ch")
+
+        self.gridLayout_4.addWidget(self.label_plot_ch, 2, 2, 1, 1)
 
         self.comboBox_plot_channel = QComboBox(self.tab_preview)
         self.comboBox_plot_channel.addItem("")
@@ -107,6 +97,30 @@ class Ui_MainWindowDesign(object):
 
         self.gridLayout_4.addWidget(self.comboBox_plot_channel, 2, 3, 1, 1)
 
+        self.label_60 = QLabel(self.tab_preview)
+        self.label_60.setObjectName(u"label_60")
+
+        self.gridLayout_4.addWidget(self.label_60, 0, 2, 1, 1)
+
+        self.checkBox_lockMove = QCheckBox(self.tab_preview)
+        self.checkBox_lockMove.setObjectName(u"checkBox_lockMove")
+
+        self.gridLayout_4.addWidget(self.checkBox_lockMove, 0, 0, 1, 1)
+
+        self.horizontalSpacer_13 = QSpacerItem(40, 20, QSizePolicy.Expanding, QSizePolicy.Minimum)
+
+        self.gridLayout_4.addItem(self.horizontalSpacer_13, 0, 4, 1, 1)
+
+        self.label_106 = QLabel(self.tab_preview)
+        self.label_106.setObjectName(u"label_106")
+
+        self.gridLayout_4.addWidget(self.label_106, 0, 5, 1, 1)
+
[…]
