### 2024-09-27 d2a1422
Update paper.md
