+++ paper.md
@@ -63,7 +63,19 @@ Fluorescence microscopy is an essential workhorse of biomedical sciences, thanks
 
 The capabilities of ISM can be further extended if coupled with a fast detector, such as an array of single photon avalanche diodes (SPAD)​ [@Buttafava1; @Slenders1]​. This type of detector features the readout of each pixel is independent and async, preserving the temporal information of each photon arrival. In other words, it is possible to tag each photon with its arrival time with respect to the excitation laser, expanding the dataset to a 5D array. This data can be used for estimating the fluorescence lifetime allowing the fluorescent lifetime imaging scanning microscopy (FLISM), a microscopy technique that merges the temporal information with the ISM advantages mentioned above​ [@Tortarolo2; @Rossetta1]​.
 
-In live-cell microscopy, the demand for high-resolution imaging coexists with the necessity to protect sample integrity. Fast pixel dwell times are instrumental in this pursuit, minimizing sample damage while enhancing image quality. Leveraging the capabilities of the SPAD array detector, which can achieve megahertz photon flux per single pixel, requires an advanced data acquisition and control system based on FPGA technology. Existing open-source microscope control software, such as ImSwitch ​[@Moreno1]​ and µManager​ [@Edelstein1]​, have been designed for camera-based instruments. Integrating an imaging-scanning microscope's control system, which supports multichannel scanning microscopy, ensures the synchronization of acquisition with a pixel dwell time as low as 1 µs, and manages high photon flux reaching megahertz rates, would compromise the performance of the aforementioned software architectures. Thus, those solutions are not practical for ISM instruments, necessitating tailored software. In this context, the BrightEye-MCS (Microscope Control Suite) is the new open-source tool written in Python, which allows the user to control a laser-scanning microscope equipped with a SPAD detector, providing a user-friendly GUI, real-time preview, and data saving in HDF5 ready to be processed internally or with external tools.
+## Statement of need
+
+In live-cell microscopy, the demand for high-resolution imaging coexists with the necessity to protect sample integrity. Fast pixel dwell times are instrumental in this pursuit, minimizing sample damage while enhancing image quality. Leveraging the capabilities of the SPAD array detector, which can achieve megahertz photon flux per single pixel, requires an advanced data acquisition and control system based on FPGA technology. Existing open-source microscope control software, such as ImSwitch ​[@Moreno1]​ and µManager​ [@Edelstein1]​, have been designed for camera-based instruments. While these tools are general-purpose, flexible, and scriptable, they are not optimized to handle the synchronization needs of ISM instruments, particularly with pixel dwell times as low as 1 µs and photon flux rates reaching megahertz levels. 
+
+In contrast, BrightEyes-MCS (Microscope Control Suite) is an open-source software specifically designed for controlling laser-scanning microscopes equipped with SPAD detectors. It overcomes the limitations of other software by providing real-time synchronization, high-throughput data acquisition, and seamless compatibility with the high-speed scanning demands of ISM and FLISM systems.
+
+BrightEyes-MCS not only features a user-friendly graphical interface (GUI) for microscope control, but also supports real-time previews and efficient data saving in the HDF5 format, which is ready for internal processing or integration with external analysis tools. Although tailored for ISM, BrightEyes-MCS is versatile and can be adapted for various applications, including fluorescence correlation spectroscopy (FCS), spectroscopy, and other advanced imaging techniques. This flexibility makes it a valuable resource for researchers working with diverse microscopy setups that require precise control and high-performance data acquisition.
+
+
+---
+
+
[…]
