### 2024-09-27 5296686
GUI improved FCS Batch tab
