### 2024-09-27 057948a
Added test for controlling the brighteyes-mcs cython libraries
