+++ test/check_cython_lib.py
@@ -4,14 +4,25 @@ import sys
 # insert root directory into python module search path
 sys.path.insert(1, os.getcwd())
 
-from brighteyes_mcs.libs.cython import fastconverter as fc
 
-print("\n\n\n========================================================================")
-print(" Try to import brighteyes_mcs.libs.cython.fastconverter")
-print(" Result: ",
-      "OK INSTALLED" if fc.convertRawDataToCountsDirect.__name__ == 'convertRawDataToCountsDirect' else "NOT INSTALLED!")
-print("========================================================================\n\n\n")
+from brighteyes_mcs.libs.cython.fastconverter import (
+      convertRawDataToCountsDirect,
+      convertRawDataToCountsDirect49,
+      convertDataFromAnalogFIFO,
+)
+from brighteyes_mcs.libs.cython.autocorrelator import Autocorrelator
+from brighteyes_mcs.libs.cython.timeBinner import timeBinner
+
 
 def test():
-      a=fc
-      pass
+      print("\n\n\n========================================================================")
+      print(" Try to import brighteyes_mcs.libs.cython.fastconverter ... ",
+            "OK!" if convertRawDataToCountsDirect.__name__ == 'convertRawDataToCountsDirect' else "FAILED!")
+      print(" Try to import brighteyes_mcs.libs.cython.autocorrelator ... ",
+            "OK!" if Autocorrelator.__name__ == 'Autocorrelator' else "FAILED!")
+      print(" Try to import brighteyes_mcs.libs.cython.timeBinner ... ",
+            "OK!" if timeBinner.__name__ == 'timeBinner' else "FAILED!")
+      print("========================================================================\n\n\n")
+
+if __name__ == '__main__':
+      test()
