+++ brighteyes_mcs/gui/main_window_design.py
@@ -2848,8 +2848,8 @@ class Ui_MainWindowDesign(object):
         self.checkBox_lockMove.setToolTip(QCoreApplication.translate("MainWindowDesign", u"<html><head/><body><p>It is locking the autorange of the image.</p></body></html>", None))
 #endif // QT_CONFIG(tooltip)
         self.checkBox_lockMove.setText(QCoreApplication.translate("MainWindowDesign", u"Lock Move", None))
-        self.label_106.setText(QCoreApplication.translate("MainWindowDesign", u"Click: Move to position", None))
-        self.label_107.setText(QCoreApplication.translate("MainWindowDesign", u"Double-Click: set a Marker", None))
+        self.label_106.setText(QCoreApplication.translate("MainWindowDesign", u"Double-Click: move to position", None))
+        self.label_107.setText(QCoreApplication.translate("MainWindowDesign", u"Ctrl+Double-Click: set a Marker", None))
         self.tabWidget.setTabText(self.tabWidget.indexOf(self.tab_preview), QCoreApplication.translate("MainWindowDesign", u"Preview", None))
         self.pushButton_13.setText(QCoreApplication.translate("MainWindowDesign", u"Move to selected offset", None))
         self.pushButton_copyPositionsMarkersFCS.setText(QCoreApplication.translate("MainWindowDesign", u"Add Markers in Batch acq. (FCS)", None))
