### 2024-09-27 #7125 @VicidominiLab
Dear @adamltyson,

I have updated the paper and the repository. Specifically, I revised the installation scripts to facilitate installation of BrightEyes-MCS for users working in a Conda environment. Additionally, I modified the README.md file to clarify the installation requirements and instructions.  In the paper, I enhanced the introduction by adding a "Statement of Need" section, emphasizing how BrightEyes-MCS differs from ImSwitch and µManager, and highlighting its necessity for ISM. And, I fixed the errors in the bibliography.

As I am not familiar with the review process in JOSS, I would appreciate it if you could let me know how to proceed.

Thank you in advance,
Mattia
