### 2024-09-30 #7125 @adamltyson
Hi @VicidominiLab,

Thanks for making these changes. The next step would be to check in with @raacampbell to make sure that with these changes, he feels that the paper is ready to accept.

However, as @beniroquai's review isn't finished yet, I would wait until this is in to see if anything else needs doing first.
