### 2024-09-30 #7125 @jacopoabramo
@VicidominiLab and @adamltyson sorry it took me so much time to write an initial review; I took the chance to follow the progression of @raacampbell's review in order to see if other points were missing beyond his. After the recent changes in the paper's structure and grammar I find the paper to be readable for my standards.

I have a remark concerning the box I checked

### Functionality

  - I can confirm that the functional claims are respected because I have the hardware in my laboratory that allows me to do so; nevertheless - as stated before - an hardware simulator to verify the functionalities of the software should be added;

Then remarks for the boxes I did not check:

### Documentation

   - other than the existing video demonstrating the functionalities in the main README page I do not see other real-life examples provided by other documentation pages; some documentation to provide an initial user-manual of how an experiment would proceed in the context of Brighteyes-MCS would greatly help newcomers;
  - as it stands, from what I can comprehend Brighteyes-MCS is presented as a stand-alone project that could be, in theory, combined with other existing solutions (as also stated in the Conclusions paragraph of the paper); the lack of an API listing or in general a detailed internal documentation of the software architecture makes this point incomplete in my opinion;
- [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
  - I could not find automated tests in the repository other the one controlling the cython libraries;
  - These are currently not available in the main README page and I could not find anywhere where these were mentioned.
