### 2024-09-30 51ce8c0
Added Contributing and Community Guidelines
