+++ README.md
@@ -2,328 +2,17 @@
 
 BrightEyes-MCS, a Microscope Control Software for image-scanning microscopy designed by the Molecular Microscopy and Spectroscopy group of the Italian Institute of Technology. 
 
+For complete documentation, please visit our [Wiki page](https://github.com/VicidominiLab/BrightEyes-MCS/wiki).
+
 [docs_video_brighteyes-mcs.webm](https://user-images.githubusercontent.com/61466143/202733339-2524c826-74d9-4ebc-8885-56855706200f.webm)
 
 <!-- [![video session](docs/video/brighteyes-frame.png)](https://user-images.githubusercontent.com/61466143/202123174-e9019c5c-bc9c-403d-b710-0516af8346b9.webm) -->
 **Click on the video**
-## Development Notice
 
-**Important:** This software is currently under active development and may contain bugs or incomplete features. Please use it with caution and report any issues you encounter to help us improve the application.
+## Quick Start
 
-Main Features
---------
+To get started, refer to the detailed instructions and guidelines available on the [Wiki](https://github.com/VicidominiLab/BrightEyes-MCS/wiki).
 
-- Realtime preview
-- Support up 25 digital channels (SPAD array) + 2 "extra" channel
-- Support up 2 analog channels (PMT)
-- Pan & Zoom in realtime
-- Scan along each axis and XYZ multi-stack
-  - (Supported up to 3 linear voltage-controlled actuator *i.e.* galvo mirror or piezo-stage)
-- Time lapse / Macro
-- Data saved in HDF5:
-  - Multi-dimensional `[repetition, z, y, x, t_bin, channel]` array.
-  - Metadata (current configurations, extra information, users comments etc etc).
-  - The data can be analyzed a visualized directly with Napari (through [Napari-ISM](https://www.napari-hub.org/plugins/napari-ISM) plugin) just by a few click
-- The pixel is subdivided in temporal bins:
-  - Normal acquisition down-to:
-    - 0.5 us for USB hardware
-    - 0.25 us for Chassis Thunderbolt configuration
-  - Histogram TCSPC, via Digital Frequency Domain techniques
-      - 0.4 ns for bins
-- Integrated Jupyter Python Console for inspecting or acting on the running program
-- Plugins / Scripting in Python
-  - Automatic grid analysis for calibration
-- Fluorescence correlation spectroscopy Preview
-- Integrated with BrightEyes-Time Tagging Module
-
-
-
-Moreover, it can be integrated with the
-[BrightEyes Time-Tagging Module](https://github.com/VicidominiLab/BrightEyes-TTM). This module allow to perform 
-single-photon time-tagging microscopy therefore fluorescence spectroscopy, fluorescence lifetime imaging microscopy
-(FLIM), and fluorescence lifetime correlation spectroscopy (FLFS) experiments.
-
----
-
-![image](docs/img/BrightEyes-NI.png)
-
----
-
-
----
-
-# Requirements 
-
-## Software
-- Windows OS
-- Python **3.10**
-- - Download [installer](https://www.python.org/ftp/python/3.10.8/python-3.10.8-amd64.exe)
-- - Or `winget install -e --id Python.Python.3.10`
-- Git
-- - [Git for Windows] (https://gitforwindows.org/)
-- - Or `winget install -e --id Git.Git`
-- FPGA drivers installed,
-- - NI R Series Multifunction [RIO driver](https://www.ni.com/en/support/downloads/drivers/download.ni-r-series-multifunction-rio.html)
-- A **C compiler** installer, either
-  - GCC [MSYS2](https://www.msys2.org/) (`winget install -e --id MSYS2.MSYS2`)
-  - or Microsoft Visual Studio with C++ build tools.
-
-
-## Hardware
[…]
