### 2024-10-01 #7125 @jacopoabramo
Community guidelines have been updated, so that's another check completed.
