### 2024-10-01 #7125 @jacopoabramo
@mattia-donato in the wiki section "Example usage" I would suggest to add small screenshots of the discussed widget in each paragraph to give a clear understanding to the user where and what to click
