### 2024-10-01 950e30a
Update paper.bib
