+++ paper.bib
@@ -72,6 +72,18 @@
   url       = {https://www.nature.com/articles/s41566-023-01112-y}
 }
 
+@article{Zunino3,
+      title={Structured Detection for Simultaneous Super-Resolution and Optical Sectioning in Laser Scanning Microscopy}, 
+      author={A. Zunino and G. Garrè and E. Perego and S. Zappone and M. Donato and G. Vicidomini},
+      year={2024},
+      eprint={2406.12542},
+      archivePrefix={arXiv},
+      primaryClass={physics.optics},
+      url={https://arxiv.org/abs/2406.12542}, 
+      doi={10.48550/arXiv.2406.12542},
+}
+
+
 @article{Buttafava1,
   author    = {M. Buttafava and F. Villa and M. Castello and G. Tortarolo and E. Conca and M. Sanzaro and S. Piazza and P. Bianchini and A. Diaspro and F. Zappa and G. Vicidomini and A. Tosi},
   title     = {SPAD-based asynchronous-readout array detectors for image-scanning microscopy},
@@ -165,4 +177,3 @@
   url          = {https://theses.hal.science/tel-04341599},
   nnt          = {2023BORD0230},
 }
-
