### 2024-10-01 307238e
Update paper.md
