### 2024-10-01 edbdafd
Added some documentation on the main parts
