+++ brighteyes_mcs/gui/main_window.py
@@ -91,10 +91,23 @@ pg.setConfigOption("foreground", "w")
 
 
 class PluginSignals(QObject):
+    """
+    Class for defining custom signals for the plugins
+    """
     signal = Signal(str)
 
 
 class MainWindow(QMainWindow):
+    """
+    Main window class for the BrightEyes-MCS application.
+
+    Attributes:
+        http_server_thread (FastAPIServerThread): Thread for the HTTP server.
+        guiReadyFlag (bool): Flag indicating if the GUI is ready.
+        init_ready (bool): Flag indicating if the initialization is ready.
+        splash (QSplashScreen): Splash screen for the application.
+    """
+
     def __init__(self, args=None):
         self.http_server_thread = None
         self.guiReadyFlag = False
@@ -488,11 +501,17 @@ class MainWindow(QMainWindow):
 
     @Slot()
     def httpServerCheckBoxChanged(self):
+        """
+        Slot for the HTTP server checkbox changed event
+        """
         if self.ui.checkBox_httpServer.isChecked():
             self.httpApiServer_start()
         else:
             self.httpApiServer_stop()
     def httpApiServer_start(self):
+        """
+        Starts the FastAPI HTTP server
+        """
         print_dec("httpApiServer_start()")
         if self.http_server_thread is None:
             self.http_server_thread = FastAPIServerThread(self, self.ui.lineEdit_httpAddr.text(), int(self.ui.lineEdit_httpPort.text()))
@@ -504,8 +523,11 @@ class MainWindow(QMainWindow):
                                             self.ui.lineEdit_httpAddr.text(), self.ui.lineEdit_httpPort.text()))
         else:
             print_dec("HTTP Server FastAPIServerThread ALREADY RUNNING")
-            
+
     def httpApiServer_stop(self):
+        """
+        Stops the FastAPI HTTP server
+        """
         print_dec("htttApiServer_stop()")
         self.http_server_thread.stop()
         print_dec("HTTP Server FastAPIServerThread Stop")
@@ -514,6 +536,10 @@ class MainWindow(QMainWindow):
         self.http_server_thread=None
 
     def configuration_helper_init(self):
+        """
+        Return a dictionary with the configuration helper for the GUI elements
+        They are used both for configuration saving and loading, and macros.
+        """
         configuration_helper = {}
 
         configuration_helper["fcs"] = (
@@ -907,6 +933,9 @@ class MainWindow(QMainWindow):
         return configuration_helper
 
     def setupAnalogOutputGUI(self):
+        """
+        Build the Analog Output GUI menu
+        """
         self.ui.spinBox_AnalogOut.deleteLater()
         self.ui.comboBox_AnalogOut.deleteLater()
         self.ui.label_AnalogOut.deleteLater()
@@ -969,6 +998,9 @@ class MainWindow(QMainWindow):
 
     @Slot()
     def analogOutChanged(self):
+        """
+        Slot for the Analog Output changed event
+        """
         print_dec("analogOutChanged()")
 
         mydict = {}
@@ -984,6 +1016,9 @@ class MainWindow(QMainWindow):
 
     @Slot()
     def laserChanged(self):
+        """
+        Slot for the Laser changed event
+        """
         print_dec("laserChanged")
         self.setRegistersDict(
             {
@@ -996,6 +1031,9 @@ class MainWindow(QMainWindow):
 
     @Slot()
     def table_keyPressEvent(self, event):
+        """
+        Slot for the table key pressed
+        """
         widget = self.ui.tableWidget
         print_dec(
             "Remove from",
@@ -1011,16 +1049,28 @@ class MainWindow(QMainWindow):
 
[…]
