+++ brighteyes_mcs/libs/spad_fcs_manager.py
@@ -12,7 +12,59 @@ from ..libs.mp_shared_array import MemorySharedNumpyArray
 
 
 class SpadFcsManager(QObject):
+    """
+    Manages the SPAD FCS (Single-Photon Avalanche Diode Fluorescence Correlation Spectroscopy) operations, including FPGA connections, data acquisition, and processing.
+
+    Attributes:
+        channels (int): Number of channels.
+        dim_detector (int): Dimension of the detector.
+        niAddr (str): NI address.
+        bitfile (str): Path to the bitfile.
+        timeout_fifos (float): Timeout for the FIFOs.
+        fpga_handle (FpgaHandle): Handle for the FPGA.
+        shared_arrays_ready (bool): Flag indicating if shared arrays are ready.
+        mp_manager (multiprocessing.Manager): Manager for multiprocessing.
+        ultra_dict_inst (dict): Dictionary for ultra configuration.
+        shared_dict (dict): Shared dictionary for multiprocessing.
+        default_configuration (dict): Default configuration for the registers.
+        requested_depth (int): Requested depth for acquisition.
+        actual_depth (int): Actual depth for acquisition.
+        timebins_per_pixel (int): Number of time bins per pixel.
+        time_resolution (float): Time resolution.
+        dim_x (int): Number of pixels in x dimension.
+        dim_y (int): Number of pixels in y dimension.
+        dim_z (int): Number of frames.
+        dim_rep (int): Number of repetitions.
+        previewProcess (AcquisitionLoopProcess): Process for previewing data.
+        expected_raw_data (int): Expected raw data.
+        expected_raw_data_per_frame (int): Expected raw data per frame.
+        registers_configuration (dict): Configuration of the registers.
+        shared_memory_buffer (MemorySharedNumpyArray): Shared memory buffer.
+        shared_image_xy (MemorySharedNumpyArray): Shared image in xy plane.
+        shared_fingerprint (MemorySharedNumpyArray): Shared fingerprint data.
+        shared_fingerprint_mask (MemorySharedNumpyArray): Shared fingerprint mask.
+        is_connected (bool): Flag indicating if connected to FPGA.
+        fifo_chuck_size (int): Size of the FIFO chuck.
+        acquisition_done_event (multiprocessing.Event): Event for acquisition done.
+        acquisition_almost_done_event (multiprocessing.Event): Event for acquisition almost done.
+        acquisition_run_event (multiprocessing.Event): Event for acquisition run.
+        acquisition_stop_event (multiprocessing.Event): Event for acquisition stop.
+        activate_preview_event (multiprocessing.Event): Event for activating preview.
+        autocorrelation_maxx (int): Maximum value for autocorrelation.
+        trace_bins (int): Number of trace bins.
+        trace_sample_per_bins (int): Number of samples per trace bin.
+        preview_buffer_size_in_words (int): Size of the preview buffer in words.
+        len_fifo_prebuffer (int): Length of the FIFO prebuffer.
+        activated_fifos_list (list): List of activated FIFOs.
+        DFD_Activate (bool): Flag for DFD activation.
+        snake_walk (bool): Flag for snake walk mode.
+        use_rust_fifo (bool): Flag for using Rust FIFO.
+        debug (bool): Debug flag.
+    """
     def __init__(self, filename="bitfiles/MyBitfileUSB.lvbitx", address="RIO0", channels=25):
+        """
+        Constructor of the class of SpaFCSManager
+        """
         super().__init__()
         # def __init__(self, filename="C:/Users/madonato/PycharmProjects/pyspad-fcs/bitfiles/laser_wait4.lvbitx", address="RIO0"):
         self.channels = channels
[…]

[…]
