+++ brighteyes_mcs/plugins/myplugin/load_plugin.py
@@ -1,3 +1,10 @@
+"""
+The load_plugin.py is the default loader of the plugin that can be modified in order to implement the plugin.
+Remind to add from .load_plugin import load_plugin in the __init__.py
+The parameter plugin_manager allows to QWidget to the plugins
+plugin_manager.addTab(context["widget"], "myplugin")
+"""
+
 from .plugin_form import myForm
 from PySide2.QtWidgets import QWidget
 
