+++ test/test_mainwindow.py
@@ -1,0 +1,415 @@
+import os
+import sys
+# insert root directory into python module search path
+sys.path.insert(1, os.getcwd())
+
+import unittest
+from unittest.mock import MagicMock, patch
+
+class TestMyClass(unittest.TestCase):
+
+    def temporalSettingsChanged_updates_registers(self):
+        instance = MyClass()
+        instance.ui.checkBox_circular.isChecked = MagicMock(return_value=True)
+        instance.ui.checkBox_DummyData.isChecked = MagicMock(return_value=False)
+        instance.setRegistersDict = MagicMock()
+
+        instance.temporalSettingsChanged()
+
+        instance.setRegistersDict.assert_called_with({
+            "Cx": int(Cx),
+            "#timebinsPerPixel": int(time_bin),
+            "ClockDur": int(clock_duration),
+            "WaitForLaser": int(waitForLaserInCycle),
+            "WaitAfterFrame": int(waitAfterFrame),
+            "WaitOnlyFirstTime": waitOnlyFirstTime,
+            "CircularMotionActivate": True,
+            "DummyData": False,
+        })
+
+    def trace_parameters_changed_updates_labels(self):
+        instance = MyClass()
+        instance.ui.doubleSpinBox_binsize.value = MagicMock(return_value=1.0)
+        instance.ui.spinBox_timeresolution.value = MagicMock(return_value=1)
+        instance.ui.doubleSpinBox_maxlength.value = MagicMock(return_value=1.0)
+        instance.ui.label_trace_total_bins.setText = MagicMock()
+        instance.ui.label_trace_total_bins.setStyleSheet = MagicMock()
+
+        instance.trace_parameters_changed()
+
+        instance.ui.label_trace_total_bins.setText.assert_called_with("1000")
+        instance.ui.label_trace_total_bins.setStyleSheet.assert_called_with("")
+
+    def configure_analog_sets_registers(self):
+        instance = MyClass()
+        instance.ui.checkBox_analog_in_integrate_AI0.isChecked = MagicMock(return_value=True)
+        instance.ui.checkBox_analog_in_invert_AI0.isChecked = MagicMock(return_value=False)
+        instance.ui.comboBox_analogSelect_A.currentIndex = MagicMock(return_value=1)
+        instance.ui.comboBox_analogSelect_B.currentIndex = MagicMock(return_value=2)
+        instance.setRegistersDict = MagicMock()
+
+        instance.configure_analog()
+
+        instance.setRegistersDict.assert_called_with({
+            "AnalogA0 integrate": True,
+            "AnalogA0 invert": False,
+            "AnalogInputA": 1,
+            "AnalogInputB": 2,
+        })
+
+    def activateShowPreview_calls_activateShowPreview(self):
+        instance = MyClass()
+        instance.spadfcsmanager_inst.activateShowPreview = MagicMock()
+
+        instance.activateShowPreview(True)
+
+        instance.spadfcsmanager_inst.activateShowPreview.assert_called_with(True)
+
+    def activateFIFOflag_sets_fifo_flags(self):
+        instance = MyClass()
+        instance.ui.radioButton_digital.isChecked = MagicMock(return_value=True)
+        instance.ui.radioButton_analog.isChecked = MagicMock(return_value=False)
+        instance.spadfcsmanager_inst.setActivatedFifo = MagicMock()
+        instance.setRegistersDict = MagicMock()
+
+        instance.activateFIFOflag()
+
+        instance.spadfcsmanager_inst.setActivatedFifo.assert_called_with(["FIFO"])
+        instance.setRegistersDict.assert_called_with({
+            "DFD_Activate": False,
+            "activateFIFOAnalog": False,
[…]
