+++ scripts/shift_vectors.py
@@ -1,3 +1,8 @@
+"""
+Script that can be launched directly on the ScriptLauncher.
+This script calculate and shows the shift vector of the last ascquired image.
+"""
+
 import brighteyes_ism.dataio.mcs as mcs
 import brighteyes_ism.analysis.APR_lib as apr
 import brighteyes_ism.analysis.Tools_lib as tools
