### 2024-10-02 #7125 @mattia-donato
Dear @jacopoabramo
Finally, I added the in the Wiki the screenshots the example of usage and I added the API.
Moreover, I added some automated tests of the core of BrightEyes-MCS.
