### 2024-10-02 #7125 @jacopoabramo
Just checked, thanks for the update @mattia-donato

@adamltyson for clarification, when writing "automated tests", would that include GitHub actions? Otherwise my review is complete - unless I have to do something else
