### 2024-10-03 a23ec32
Updated firmware downloader in the installer 2
Updated firmware downloader in the installer

Updated firmware downloader in the installer

Added TTM GUI
