+++ brighteyes_mcs/gui/ttm_widget.py
@@ -1,0 +1,924 @@
+from PySide2.QtWidgets import (
+    QWidget,
+    QFileDialog,
+)  # QDoubleSpinBox, QComboBox, QCheckBox,
+
+from PySide2.QtCore import Slot, Qt, QAbstractTableModel  # , QUrl, QObject,
+from PySide2.QtGui import QFont
+import numpy as np
+
+import matplotlib.pyplot as plt
+
+from matplotlib.backends.backend_qtagg import (
+    FigureCanvas,
+)
+from matplotlib.figure import Figure
+
+
+#
+# import shutil
+# import h5py
+#
+# from ..libs.multithread_caller_helper import pushButton_multithread_call
+#
+# from PySide2.QtCore import (
+#     QThreadPool,
+# )  # , Slot, Signal, SIGNAL, QTimer, Qt, QDir, QFile, QCoreApplication, QTime, QRunnable
+
+
+#
+# import numpy as np
+# import matplotlib.pyplot as plt
+# from matplotlib.colors import hsv_to_rgb
+#
+# from matplotlib.backends.backend_qtagg import (
+#     FigureCanvas,
+# )
+# from matplotlib.figure import Figure
+
+# from brighteyes_mcs.gui import ttm_widget_design
+import ttm_widget_design
+import ttm_recv
+
+
+class TtmWidget(QWidget, ttm_widget_design.Ui_Form):
+    def __init__(self, main_window=None, channels=25):
+        super().__init__()
+        self.setupUi(self)
+
+        self.pushButton_start.clicked.connect(self.cmd_start)
+        self.pushButton_stop.clicked.connect(self.cmd_stop)
+        self.pushButton_update.clicked.connect(self.cmd_update)
+
+        self.canvas_hist = FigureCanvas(Figure(figsize=(5, 7)))
+        self.tab_hist = self.gridLayout_central.addWidget(self.canvas_hist)
+        self.multi_ax_hist = self.canvas_hist.figure.subplots(2, 2)
+
+        self.ax_img = self.multi_ax_hist[0, 0]
+        self.ax_fingerprint = self.multi_ax_hist[0, 1]
+        self.canvas_hist.draw()
+
+        self.im_img = self.ax_img.imshow([[0,0],[1,1]])
+        self.im_fingerprint = self.ax_fingerprint.imshow([[0,0],[1,1]])
+
+        # plt.colorbar(self.im_img, cax=self.multi_ax_hist[1, 0], orientation='vertical')
+        # plt.colorbar(self.im_fingerprint, cax=self.multi_ax_hist[1, 1], orientation='vertical')
+
+        self.ttmrecv = ttm_recv.TtmRecv()
+        self.ttmrecv.init_threads()
+        self.ttmrecv.cmd_do_not_save()
+        self.cmd_update()
+
+
+    @Slot()
+    def cmd_start(self):
+        self.ttmrecv.cmd_start()
+
+
+        print("cmd_start")
+
+    @Slot()
+    def cmd_stop(self):
+        self.ttmrecv.cmd_stop()
+        self.ttmrecv = ttm_recv.TtmRecv()
+        print("cmd_stop")
+
+    @Slot()
+    def cmd_update(self):
+        print("cmd_update")
+        s = self.ttmrecv.status()
+        p = self.ttmrecv.get_preview()
+        self.plainTextEdit_debug.setPlainText(str(s))
+        d = np.asarray(s["count_photons"])
+        x_max = s["count_pixel_max"]
+        y_max = s["count_line_max"]
+
+        d = self.reorder_25ch(d)[0:25].reshape(5,5)
+        self.ax_fingerprint.imshow(d)
+        if not (x_max==0 or y_max==0):
[…]
