### 2024-10-03 #7125 @adamltyson
Hi @jacopoabramo,

Although CI would be ideal, it is not necessary for acceptance. The JOSS criteria for tests are as follows:

>Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

>    Good: An automated test suite hooked up to continuous integration (GitHub Actions, Circle CI, or similar)
>    OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., >a sample input file to assert behavior)
>    Bad (not acceptable): No way for you, the reviewer, to objectively assess whether the software works

Although we will always aim for the "Good", the only criteria for acceptance is the "OK".
