### 2024-10-04 d5cb0e7
Updated firmware downloader in the installer
