+++ download_firmware.py
@@ -1,15 +1,75 @@
-from PySide2.QtCore import Qt, QUrl
-from PySide2.QtWidgets import QMainWindow, QApplication, QPushButton, QWidget, QVBoxLayout, QHBoxLayout, QFrame, QLabel, QMessageBox
-from PySide2.QtWebEngineWidgets import QWebEngineView
-
 import sys, os
-
 import requests, zipfile
 from io import BytesIO
+import markdown
+from PySide2.QtCore import Qt, QThread, Signal
+from PySide2.QtWidgets import QMainWindow, QApplication, QPushButton, QWidget, QVBoxLayout, QHBoxLayout, QFrame, QLabel, QMessageBox, QTextBrowser, QProgressDialog, QDialog, QProgressBar, QVBoxLayout
 
-# Define the URL of the ZIP file you want to download
+# Define the URL of the Markdown file you want to download
+md_url = "https://raw.githubusercontent.com/VicidominiLab/BrightEyes-MCSLL/refs/heads/main/LICENSE.md"
 zip_url = "https://github.com/VicidominiLab/BrightEyes-MCSLL/archive/refs/heads/main.zip"
 extract_directory = "brighteyes_mcs/bitfiles"
+
+class DownloadThread(QThread):
+    download_progress = Signal(int)
+
+    def __init__(self, url):
+        super().__init__()
+        self.url = url
+        self.response = None
+        self.total_size = 0
+
+    def run(self):
+        try:
+            # Start streaming the download
+            self.response = requests.get(self.url, stream=True)
+            self.response.raise_for_status()  # Check for HTTP errors
+            self.total_size = int(self.response.headers.get('content-length', 0))
+            print(self.response)
+            print(self.response.headers.get('content-length', 0))
+
+            downloaded_size = 0
+            chunk_size = 1024  # 1KB
+            content = BytesIO()
+
+            # Download in chunks and update progress
+            for chunk in self.response.iter_content(chunk_size=chunk_size):
+                if chunk:  # Filter out keep-alive chunks
+                    content.write(chunk)
+                    downloaded_size += len(chunk)
+                    progress_percent = int(downloaded_size / self.total_size * 100)
+                    self.download_progress.emit(progress_percent)
+
+            content.seek(0)
+            self.content = content  # Store the content for extraction later
+
+        except requests.exceptions.RequestException as e:
+            print(f"Error fetching the ZIP file: {e}")
+            error_message = (
+                "=============================================\n"
+                "ERROR: Firmware Downloading Failed\n"
+                f"Please download it manually from: {self.url}\n"
+                "============================================="
+            )
+            print(error_message)
+            msgBox = QMessageBox()
+            msgBox.setText("Firmware Downloading Failed. Check details on console.")
+            msgBox.exec_()
+
+class ProgressDialog(QDialog):
+    def __init__(self):
+        super().__init__()
+        self.setWindowTitle("Downloading Firmware")
+        self.setGeometry(300, 300, 400, 100)
+
+        layout = QVBoxLayout(self)
+
+        self.progress_bar = QProgressBar(self)
+        self.progress_bar.setMaximum(100)
+        layout.addWidget(self.progress_bar)
+
+    def set_progress(self, value):
+        self.progress_bar.setValue(value)
 
 class MainWindow(QMainWindow):
 
@@ -20,10 +80,12 @@ class MainWindow(QMainWindow):
         container = QWidget(self)
         container_layout = QVBoxLayout()
 
-        self.browser = QWebEngineView()
-        self.browser.setUrl(QUrl("https://github.com/VicidominiLab/BrightEyes-MCSLL/blob/main/LICENSE.md"))
+        # Create a QTextBrowser to display the downloaded Markdown content
[…]
