### 2024-10-04 e770beb
Updated firmware downloader in the installer
