### 2024-10-04 7c57068
Updated firmware downloader in the installer 2
