### 2024-10-04 5241de9
Updated firmware downloader in the installer
