### 2024-10-04 0befd6b
download_firmware.py updated. Now the size is requested before start the download stream
