### 2024-10-04 678b58b
download_firmware.py hope to solve the bug  'content-length'=0
