### 2024-10-04 #7125 @VicidominiLab
Hi everyone,

Just a quick update: version v0.9.2 has now been tagged and released. Many thanks to @raacampbell and @jacopoabramo for your feedback!!!

I’d like to ask @adamltyson how it works now to proceed.

Also, just to note, I’ll be out of the office next week.

Thanks again for all your support!

Bests regards,
Mattia
