### 2024-10-07 #7125 @adamltyson
Hi @raacampbell, @jacopoabramo, Can I just check that you're happy to recommend this submission for acceptance and that all of the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) have been met?

@VicidominiLab, once the reviewers have confirmed, I will do some final checks, and then there's some admin to do before I will pass this onto the associate editor in chief for final acceptance.
