### 2024-10-07 VicidominiLab/BrightEyes-MCS#15 @jacopoabramo
Automating the switch to a `conda` or `mamba` environment
In some cases, more often than not, scientific users will tend to use `conda` virtual environments (in my case, I use `mamba` environments for example) to install a python software package. I would suggest to automatically switch to check if the package is currently being installed in such environments and avoid spawning a second nested `venv` when this happens. It is non-critical and the software still works fine, but it helps to reduce unnecessary folder nesting and renders the overall occupied disk space reduced.

I'm leaving this issue here but it is not inherently an "issue", more like a future enhancement suggestion.
