### 2024-10-07 #7125 @jacopoabramo
Hi @adamltyson; after reviewing both the paper and the code repository I give the green light: as far as I can tell the review guidelines criteria have been met on my part and I marked all the checklist points required for the review.

There are some minor considerations to take into account to render the software package more robust, listed as follows:

- integrating a CI workflow for test automation (I believe that after what it is written in the review guidelines this is still not an hinderance for allowing publication);
- automate the switch to a `conda` or `mamba` based environment (I addressed this in issue [#15](https://github.com/VicidominiLab/BrightEyes-MCS/issues/15))

Having said that, I **reccomend** this submission for acceptance.
