### 2024-10-07 #7125 @raacampbell
> Hi @raacampbell, @jacopoabramo, Can I just check that you're happy to recommend this submission for acceptance and that all of the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) have been met?
👍
