### 2024-10-15 bcaed2e
Update paper.md
