### 2024-10-28 #7125 @adamltyson
Hi @VicidominiLab, as @beniroquai isn't able to complete their review, I think we can proceed with the two we have.

A few comments higher up, there is a post including "Additional Author Tasks After Review is Complete". Could you make sure to go through each of these tasks, and then we can proceed.

The paper is looking good, the only tiny typographical errors I noticed were a few missing spaces immediately before the in-text citations. Could you correct these?
