### 2024-10-30 37dd0f4
Update paper.md
Paper.md added spaces
