+++ paper.md
@@ -47,7 +47,7 @@ affiliations:
 #   index: 2
 # - name: Independent Researcher, Country
 #   index: 3
-date: 13 August 2017
+date: 30 October 2024
 bibliography: paper.bib
 
 # Optional fields if submitting to a AAS journal too, see this blog post:
@@ -61,7 +61,7 @@ bibliography: paper.bib
 
 Fluorescence microscopy is an essential workhorse of biomedical sciences, thanks to its capability to provide specific and quantitative information on the observed specimens. The advent of super-resolution techniques has further improved the quality of the images produced by optical microscopes. Among the numerous super-resolution techniques, image scanning microscopy (ISM) emerged as a robust and reliable technique, being able to provide gentle imaging at a high signal-to-noise ratio (SNR) with excellent optical sectioning​ [@Castello1; @Tortarolo1; @Perego1]​. An ISM microscope shares the same architecture as a confocal laser scanning microscope (CLSM) but exploits a pixeled detector. Each detector element acts as a displaced small pinhole and generates a confocal-like image after a full scan of the field of view. As the detector is made by a matrix of pixels, the raw ISM dataset is not a 2D image, it is a 4D dataset, which can be seen in two ways: each single pixel scanned on the sample plane is associated with a micro-image of the detector; vice-versa, for each physical pixel on the detector plane is associated a full-scanned field of view on the sample plane [@Sheppard1; @Muller1]​. Using tailored reconstruction algorithms, the images of the raw ISM dataset can be fused to produce a single super-resolution image using all the photons collected by the detector, guaranteeing an excellent SNR​ [@Zunino1; @Zunino2; @Zunino3]​.
 
-The capabilities of ISM can be further extended if coupled with a fast detector, such as an array of single photon avalanche diodes (SPAD)​ [@Buttafava1; @Slenders1]​. This type of detector features the readout of each pixel is independent and async, preserving the temporal information of each photon arrival. In other words, it is possible to tag each photon with its arrival time with respect to the excitation laser, expanding the dataset to a 5D array. This data can be used for estimating the fluorescence lifetime allowing the fluorescent lifetime imaging scanning microscopy (FLISM), a microscopy technique that merges the temporal information with the ISM advantages mentioned above​ [@Tortarolo2; @Rossetta1]​.
+The capabilities of ISM can be further extended if coupled with a fast detector, such as an array of single photon avalanche diodes (SPAD)​  [@Buttafava1; @Slenders1]​. This type of detector features the readout of each pixel is independent and async, preserving the temporal information of each photon arrival. In other words, it is possible to tag each photon with its arrival time with respect to the excitation laser, expanding the dataset to a 5D array. This data can be used for estimating the fluorescence lifetime allowing the fluorescent lifetime imaging scanning microscopy (FLISM), a microscopy technique that merges the temporal information with the ISM advantages mentioned above​ [@Tortarolo2; @Rossetta1]​.
 
 ## Statement of need
 
@@ -73,7 +73,7 @@ BrightEyes-MCS not only features a user-friendly graphical interface (GUI) for m
 
 ---
 
-BrightEyes is the name of the project founded by the ERC in 2018 (Consolidator Grant, N. 818699). In this context, other open-source tools have been developed, like BrighEyes-TTM, an open-hardware time-tagging module [@Rossetta1], BrightEyes-ISM[@Zunino2], a python library for ISM data analysis.
[…]
