### 2024-10-30 556758a
Delete paper.bib
