### 2024-10-30 23b5941
Delete paper.md
