### 2024-10-30 #7125 @VicidominiLab
Dear @adamltyson ,
Finally, I've fixed the typo you metioned and finally I uploaded the repository on Zenodo.
The Zenodo DOI is the following: https://doi.org/10.5281/zenodo.14012812
The version is the "v1.0.0" in the GitHub tag and release, as well in Zenodo.
The license as been fixed to the GPLv3 in Zenodo as it is in GitHub.
Thanks
Mattia
