### 2024-11-04 #7125 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
