### 2024-11-15 #7125 @Kevin-Mattheus-Moerman
@adamltyson thanks for editing here. I am picking this up now. I think we cannot continue as it stands, and I need some clarifications.

Here are my observations (but do correct me if I missed something):
* It looks like only @jacopoabramo fully completed the review.
* The last reference to @beniroquai seems to say their review was not finished (https://github.com/openjournals/joss-reviews/issues/7125#issuecomment-2383071234).
* In https://github.com/openjournals/joss-reviews/issues/7125#issuecomment-2397235419 we see @raacampbell recommends acceptance but their checklist is not completed.

@adamltyson Can you work with @raacampbell to complete the checklist, or explain why they cannot complete it further (e.g. are they missing the required hardware/software to complete those steps?). To proceed we need two completed reviews. Any clarification here would be great.
