### 2024-11-15 #7125 @adamltyson
Hi @Kevin-Mattheus-Moerman,

I think your observations are correct. My mistake, I was proceeding based on @raacampbell's acceptance, and I had assumed the checklist was effectively complete, just not filled in.

@raacampbell are you able to complete the checklist above, or are there any blockers?
