### 2024-11-15 #7125 @raacampbell
I have not installed the software so I didn't check off those boxes. Realistically to fully check the software I need to run it on  microscope and I can't do that, so I'm leaving those boxes unchecked.
