### 2024-11-18 #7125 @raacampbell
> Thanks @raacampbell. Do you think there's anything the authors could do to help you test the functionality adequately (e.g. mock hardware)? Otherwise I think we may have to see if we can find another reviewer.

Might have to get another reviewer, as the authors replied above saying that a simulated mode is out of scope.
