### 2024-11-18 #7125 @VicidominiLab
@raacampbell @adamltyson
Thanks to the reviewers' feedback, the installation process has been significantly improved. The wiki pages now provide detailed, step-by-step instructions on how to install the software, including a link to the required third-party driver.

Additionally, I would like to highlight that **the GUI can also run on machines without the NI FPGA hardware** (as tested on Virtual Machine). However, in such cases, the acquisition and preview buttons will not function as intended. On the other hand, we believe that introducing a simulation mode is beyond the scope of this work.
