### 2024-11-18 #7125 @Kevin-Mattheus-Moerman
@adamltyson Would it be worth contacting @beniroquai (e.g. by email), to see if they are you still able to finalise the review? Or did they have the same hardware issue?
