### 2024-11-21 #7125 @adamltyson
Thanks @beniroquai, much appreciated. That tickbox states " If the paper contains no such data, please check this item.", so ywould you mind just ticking it off for completeness?

@Kevin-Mattheus-Moerman, I _think_ we're good to go now with this, once the final box is ticked.
