### 2024-11-23 f938a82
Create CONTRIBUTING.md
