+++ CONTRIBUTING.md
@@ -1,0 +1,20 @@
+# Community Guidelines
+
+We welcome contributions to BrightEyes-MCS from the community. Below are the general guidelines for those who want to get involved:
+
+1. **Contributing**: If you want to contribute to the software, you can fork the repository, make your changes, and submit a pull request. We recommend getting in touch with the maintainers before starting work on major features to avoid duplicating efforts.
+2. **Reporting Issues**: You can report any bugs, issues, or feature requests by opening a new issue on the GitHub repository. Please ensure that you include as much detail as possible, including steps to reproduce the issue and your environment details (operating system, Python version, hardware setup).
+3. **Seeking Support**: For questions or technical assistance, you can start a discussion in the "Discussions" tab of the repository, or check the FAQ section in the project's documentation. For immediate support or collaborations, you can contact the developers through the contact information provided on the project website.
+
+# Contributing
+
+Contributions are welcome in the form of code, documentation, testing, or suggestions. Here's how you can contribute:
+
+1. **Fork the Repository**: Create a personal copy of the BrightEyes-MCS repository by clicking on "Fork".
+2. **Clone the Repository**: Clone the forked repository to your local machine using `git clone`.
+3. **Create a Branch**: Create a new branch for your feature or bug fix.
+4. **Make Changes**: Implement your changes, ensuring that your code is well-documented and follows the project's coding guidelines.
+5. **Test Your Changes**: Ensure that all tests pass and add new ones if necessary.
+6. **Submit a Pull Request**: Once you're happy with your changes, submit a pull request to the main repository for review. Provide a clear description of the changes and reference any related issues.
+
+For detailed instructions on how to set up your development environment, please refer to the [Wiki](https://github.com/VicidominiLab/BrightEyes-MCS/wiki).
