+++ README.md
@@ -17,7 +17,7 @@ To get started, refer to the detailed instructions and guidelines available on t
 
 **Important:** This software is currently under active development and may contain bugs or incomplete features. Please use it with caution and report any issues you encounter to help us improve the application. 
 
-For contribution guidelines, refer to the [Community Guidelines](CONTRIBUTING.MD).
+For contribution guidelines, refer to the [contributing.md](CONTRIBUTING.MD).
 
 ## License
 License
