### 2024-11-23 #7125 @mattia-donato
Many thanks to @beniroquai !

@Kevin-Mattheus-Moerman Thank you for your feedback. The contributing guidelines were already present in the Wiki pages. To improve accessibility, I have also included them in the README.md and created a CONTRIBUTING.md file containing the same information.
